+++ packages/opentelemetry-node/types/instrumentations.d.ts
@@ -69,13 +69,9 @@ export type InstrumentaionsMap = {
  *      '@opentelemetry/instrumentation-express': {
  *          enabled: false,
  *      },
- *      // You can replace a instrumentation by using a funciton
- *      '@opentelemetry/instrumentation-mongodb': () => {
- *          return new MyMongoDBInstrumentation();
- *      }
  * });
  *
- * const sdk = new ElasticNodeSDK({
+ * startNodeSDK({
  *      instrumentations: [
  *          ...customInstrumentations,
  *          // You can add here instrumentations from other sources